added global teams dropdown
