+++ frontend/pages/hosts/ManageHostsPage/_styles.scss
@@ -157,90 +157,6 @@
     }
   }
 
-  &__team-dropdown {
-    border: 0 !important;
-    position: relative;
-
-    :hover {
-      cursor: pointer !important;
-    }
-
-    &.is-focused {
-      .Select-control {
-        border: 0 !important;
-        height: 32px;
-      }
-    }
-
-    .Select-menu-outer {
-      position: absolute;
-      width: 330px;
-      min-width: 125px;
-      left: -12px;
-      border-radius: 6px;
-    }
-
-    .Select-control {
-      background-color: #fff;
-      border: 0 !important;
-      border-radius: none;
-      position: none;
-      width: max-content; // move select arrow
-      height: 20px;
-
-      &:hover {
-        box-shadow: none;
-      }
-
-      &:hover .Select-value-label {
-        color: $core-vibrant-blue !important;
-      }
-
-      .Select-arrow-zone {
-        padding-left: $pad-small;
-
-        .Select-arrow {
-          top: 1px !important;
-          margin-top: 0 !important;
-        }
-      }
-
-      .Select-multi-value-wrapper {
-        width: max-content; // move select arrow
-        height: 20px;
-        margin-bottom: $pad-xsmall;
-
-        .Select-input {
-          display: none !important;
-        }
-
-        .Select-value {
-          position: relative; // move select arrow
-          display: inline-block; // move select arrow
-          line-height: 28px;
-          padding: 0;
-          border: 0 !important;
-          background-color: #fff !important;
-          right: 0;
-          left: 0;
-          bottom: 0;
-          top: 0;
-
-          &.is-focused {
-            border: 0 !important;
-          }
-          :hover {
-            border: 0 !important;
-          }
-
-          .Select-value-label {
-            font-size: $large !important;
-          }
-        }
-      }
-    }
-  }
-
   &__labels-policies-wrap {
     margin-bottom: $pad-medium;
   }
